新增选图大小限制逻辑

+++ easyPhotos/src/main/java/com/huantansheng/easyphotos/utils/file/FileUtils.java
@@ -376,8 +376,8 @@ public class FileUtils {
      * @return
      * @author paulburke
      */
-    public static String getReadableFileSize(int size) {
-        final int BYTES_IN_KILOBYTES = 1024;
+    public static String getReadableFileSize(long size) {
+        final long BYTES_IN_KILOBYTES = 1024;
         final DecimalFormat dec = new DecimalFormat("###.#");
         final String KILOBYTES = " KB";
         final String MEGABYTES = " MB";
